+++ player-service/src/main/java/com/g1/mychess/player/service/PlayerService.java
@@ -21,7 +21,7 @@ public interface PlayerService {
 
     void whitelistPlayer(Long playerId);
 
-    void updatePlayerRating(Long playerId, int glickoRating, double ratingDeviation, double volatility);
+    void updatePlayerRating(Long playerId, double glickoRating, double ratingDeviation, double volatility);
 
     AdminPlayerDTO getPlayerDetailsForAdmin(Long playerId);
 }
